+++ packages/api/src/index.ts
@@ -1,5 +1,4 @@
 import { IApi } from './types'
-import Account from './account'
 import Event from './event'
 import Execution from './execution'
 import Instance from './instance'
@@ -8,9 +7,8 @@ import Process from './process'
 import Service from './service'
 import Runner from './runner'
 
-export default class API implements IApi {
+class API implements IApi {
 
-  account: Account
   event: Event
   execution: Execution
   instance: Instance
@@ -20,7 +18,6 @@ export default class API implements IApi {
   runner: Runner
 
   constructor(endpoint: string) {
-    this.account = new Account(endpoint)
     this.event = new Event(endpoint)
     this.execution = new Execution(endpoint)
     this.instance = new Instance(endpoint)
@@ -30,3 +27,6 @@ export default class API implements IApi {
     this.runner = new Runner(endpoint)
   }
 } 
+
+export default API;
+(module).exports = API;