+++ packages/api/src/mock.ts
@@ -24,12 +24,6 @@ export const streams = {
 }
 
 export default (endpoint: string): IApi => ({
-  account: {
-    async get() { return {} },
-    async list() { return { accounts: [] } },
-    async create() { return { name: "name", mnemonic: "mnemonic" } },
-    async delete() { return {} },
-  },
   event: {
     async create() { return { hash } },
     stream() { return streams.event },
@@ -38,7 +32,8 @@ export default (endpoint: string): IApi => ({
     async create() { return { hash } },
     async get() { return { parentHash: hash, eventHash: Buffer.from('xxx'), status: 0, instanceHash: hash, taskKey: 'xxx', inputs: encode({}) } },
     stream() { return streams.execution },
-    async update() { return {} }
+    async update() { return {} },
+    async list() { return { executions: [] } },
   },
   instance: {
     async get() { return { serviceHash: hash } },