+++ packages/api/src/typedef/process.d.ts
@@ -16,8 +16,8 @@ declare namespace mesg {
                 /** Process hash */
                 hash?: (Uint8Array|null);
 
-                /** Process key */
-                key?: (string|null);
+                /** Process name */
+                name?: (string|null);
 
                 /** Process nodes */
                 nodes?: (mesg.types.Process.INode[]|null);
@@ -38,8 +38,8 @@ declare namespace mesg {
                 /** Process hash. */
                 public hash: Uint8Array;
 
-                /** Process key. */
-                public key: string;
+                /** Process name. */
+                public name: string;
 
                 /** Process nodes. */
                 public nodes: mesg.types.Process.INode[];
@@ -52,6 +52,9 @@ declare namespace mesg {
 
                 /** Properties of a Node. */
                 interface INode {
+
+                    /** Node key */
+                    key?: (string|null);
 
                     /** Node result */
                     result?: (mesg.types.Process.Node.IResult|null);
@@ -78,6 +81,9 @@ declare namespace mesg {
                      */
                     constructor(properties?: mesg.types.Process.INode);
 
+                    /** Node key. */
+                    public key: string;
+
                     /** Node result. */
                     public result?: (mesg.types.Process.Node.IResult|null);
 
@@ -102,9 +108,6 @@ declare namespace mesg {
                     /** Properties of a Result. */
                     interface IResult {
 
-                        /** Result key */
-                        key?: (string|null);
-
                         /** Result instanceHash */
                         instanceHash?: (Uint8Array|null);
 
@@ -121,9 +124,6 @@ declare namespace mesg {
                          */
                         constructor(properties?: mesg.types.Process.Node.IResult);
 
-                        /** Result key. */
-                        public key: string;
-
                         /** Result instanceHash. */
                         public instanceHash: Uint8Array;
 
@@ -133,9 +133,6 @@ declare namespace mesg {
 
                     /** Properties of an Event. */
                     interface IEvent {
-
-                        /** Event key */
-                        key?: (string|null);
 
                         /** Event instanceHash */
                         instanceHash?: (Uint8Array|null);
@@ -153,9 +150,6 @@ declare namespace mesg {
                          */
                         constructor(properties?: mesg.types.Process.Node.IEvent);
 
-                        /** Event key. */
-                        public key: string;
-
                         /** Event instanceHash. */
                         public instanceHash: Uint8Array;
 
@@ -165,9 +159,6 @@ declare namespace mesg {
 
                     /** Properties of a Task. */
                     interface ITask {
-
-                        /** Task key */
-                        key?: (string|null);
 
                         /** Task instanceHash */
                         instanceHash?: (Uint8Array|null);
@@ -185,9 +176,6 @@ declare namespace mesg {
                          */
                         constructor(properties?: mesg.types.Process.Node.ITask);
 
-                        /** Task key. */
-                        public key: string;
-
                         /** Task instanceHash. */
                         public instanceHash: Uint8Array;
 
@@ -198,11 +186,8 @@ declare namespace mesg {
                     /** Properties of a Map. */
                     interface IMap {
 
-                        /** Map key */
-                        key?: (string|null);
-
                         /** Map outputs */
-                        outputs?: (mesg.types.Process.Node.Map.IOutput[]|null);
+                        outputs?: ({ [k: string]: mesg.types.Process.Node.Map.IOutput }|null);
                     }
 
                     /** Represents a Map. */
@@ -214,11 +199,8 @@ declare namespace mesg {
                          */
                         constructor(properties?: mesg.types.Process.Node.IMap);
 
-                        /** Map key. */
-                        public key: string;
-
                         /** Map outputs. */
-                        public outputs: mesg.types.Process.Node.Map.IOutput[];
+                        public outputs: { [k: string]: mesg.types.Process.Node.Map.IOutput };
                     }
 
                     namespace Map {
@@ -226,14 +208,26 @@ declare namespace mesg {
                         /** Properties of an Output. */
                         interface IOutput {
 
-                            /** Output key */
-                            key?: (string|null);
+                            /** Output null */
+                            "null"?: (mesg.types.Process.Node.Map.Output.Null|null);
+
+                            /** Output stringConst */
+                            stringConst?: (string|null);
+
+                            /** Output doubleConst */
+                            doubleConst?: (number|null);
+
+                            /** Output boolConst */
+                            boolConst?: (boolean|null);
 
                             /** Output ref */
                             ref?: (mesg.types.Process.Node.Map.Output.IReference|null);
 
-                            /** Output constant */
-                            constant?: (mesg.protobuf.IValue|null);
+                            /** Output list */
+                            list?: (mesg.types.Process.Node.Map.Output.IList|null);
+
+                            /** Output map */
+                            map?: (mesg.types.Process.Node.Map.Output.IMap|null);
                         }
 
                         /** Represents an Output. */
@@ -245,20 +239,77 @@ declare namespace mesg {
                              */
                             constructor(properties?: mesg.types.Process.Node.Map.IOutput);
 
-                            /** Output key. */
-                            public key: string;
+                            /** Output null. */
+                            public null: mesg.types.Process.Node.Map.Output.Null;
+
+                            /** Output stringConst. */
+                            public stringConst: string;
+
+                            /** Output doubleConst. */
+                            public doubleConst: number;
+
+                            /** Output boolConst. */
+                            public boolConst: boolean;
 
                             /** Output ref. */
                             public ref?: (mesg.types.Process.Node.Map.Output.IReference|null);
 
-                            /** Output constant. */
-                            public constant?: (mesg.protobuf.IValue|null);
+                            /** Output list. */
+                            public list?: (mesg.types.Process.Node.Map.Output.IList|null);
+
+                            /** Output map. */
+                            public map?: (mesg.types.Process.Node.Map.Output.IMap|null);
 
                             /** Output value. */
-                            public value?: ("ref"|"constant");
+                            public value?: ("null"|"stringConst"|"doubleConst"|"boolConst"|"ref"|"list"|"map");
                         }
 
                         namespace Output {
+
+                            /** Null enum. */
+                            enum Null {
+                                NULL_VALUE = 0
+                            }
+
+                            /** Properties of a List. */
+                            interface IList {
+
+                                /** List outputs */
+                                outputs?: (mesg.types.Process.Node.Map.IOutput[]|null);
+                            }
+
+                            /** Represents a List. */
+                            class List implements IList {
+
+                                /**
+                                 * Constructs a new List.
+                                 * @param [properties] Properties to set
+                                 */
+                                constructor(properties?: mesg.types.Process.Node.Map.Output.IList);
+
+                                /** List outputs. */
+                                public outputs: mesg.types.Process.Node.Map.IOutput[];
+                            }
+
+                            /** Properties of a Map. */
+                            interface IMap {
+
+                                /** Map outputs */
+                                outputs?: ({ [k: string]: mesg.types.Process.Node.Map.IOutput }|null);
+                            }
+
+                            /** Represents a Map. */
+                            class Map implements IMap {
+
+                                /**
+                                 * Constructs a new Map.
+                                 * @param [properties] Properties to set
+                                 */
+                                constructor(properties?: mesg.types.Process.Node.Map.Output.IMap);
+
+                                /** Map outputs. */
+                                public outputs: { [k: string]: mesg.types.Process.Node.Map.IOutput };
+                            }
 
                             /** Properties of a Reference. */
                             interface IReference {
@@ -266,8 +317,8 @@ declare namespace mesg {
                                 /** Reference nodeKey */
                                 nodeKey?: (string|null);
 
-                                /** Reference key */
-                                key?: (string|null);
+                                /** Reference path */
+                                path?: (mesg.types.Process.Node.Map.Output.Reference.IPath|null);
                             }
 
                             /** Represents a Reference. */
@@ -282,17 +333,52 @@ declare namespace mesg {
                                 /** Reference nodeKey. */
                                 public nodeKey: string;
 
-                                /** Reference key. */
-                                public key: string;
+                                /** Reference path. */
+                                public path?: (mesg.types.Process.Node.Map.Output.Reference.IPath|null);
+                            }
+
+                            namespace Reference {
+
+                                /** Properties of a Path. */
+                                interface IPath {
+
+                                    /** Path key */
+                                    key?: (string|null);
+
+                                    /** Path index */
+                                    index?: (number|Long|null);
+
+                                    /** Path path */
+                                    path?: (mesg.types.Process.Node.Map.Output.Reference.IPath|null);
+                                }
+
+                                /** Represents a Path. */
+                                class Path implements IPath {
+
+                                    /**
+                                     * Constructs a new Path.
+                                     * @param [properties] Properties to set
+                                     */
+                                    constructor(properties?: mesg.types.Process.Node.Map.Output.Reference.IPath);
+
+                                    /** Path key. */
+                                    public key: string;
+
+                                    /** Path index. */
+                                    public index: (number|Long);
+
+                                    /** Path path. */
+                                    public path?: (mesg.types.Process.Node.Map.Output.Reference.IPath|null);
+
+                                    /** Path selector. */
+                                    public selector?: ("key"|"index");
+                                }
                             }
                         }
                     }
 
                     /** Properties of a Filter. */
                     interface IFilter {
-
-                        /** Filter key */
-                        key?: (string|null);
 
                         /** Filter conditions */
                         conditions?: (mesg.types.Process.Node.Filter.ICondition[]|null);
@@ -306,9 +392,6 @@ declare namespace mesg {
                          * @param [properties] Properties to set
                          */
                         constructor(properties?: mesg.types.Process.Node.IFilter);
-
-                        /** Filter key. */
-                        public key: string;
 
                         /** Filter conditions. */
                         public conditions: mesg.types.Process.Node.Filter.ICondition[];
@@ -384,108 +467,6 @@ declare namespace mesg {
                     /** Edge dst. */
                     public dst: string;
                 }
-            }
-        }
-
-        /** Namespace protobuf. */
-        namespace protobuf {
-
-            /** Properties of a Struct. */
-            interface IStruct {
-
-                /** Struct fields */
-                fields?: ({ [k: string]: mesg.protobuf.IValue }|null);
-            }
-
-            /** Represents a Struct. */
-            class Struct implements IStruct {
-
-                /**
-                 * Constructs a new Struct.
-                 * @param [properties] Properties to set
-                 */
-                constructor(properties?: mesg.protobuf.IStruct);
-
-                /** Struct fields. */
-                public fields: { [k: string]: mesg.protobuf.IValue };
-            }
-
-            /** Properties of a Value. */
-            interface IValue {
-
-                /** Value nullValue */
-                nullValue?: (mesg.protobuf.NullValue|null);
-
-                /** Value numberValue */
-                numberValue?: (number|null);
-
-                /** Value stringValue */
-                stringValue?: (string|null);
-
-                /** Value boolValue */
-                boolValue?: (boolean|null);
-
-                /** Value structValue */
-                structValue?: (mesg.protobuf.IStruct|null);
-
-                /** Value listValue */
-                listValue?: (mesg.protobuf.IListValue|null);
-            }
-
-            /** Represents a Value. */
-            class Value implements IValue {
-
-                /**
-                 * Constructs a new Value.
-                 * @param [properties] Properties to set
-                 */
-                constructor(properties?: mesg.protobuf.IValue);
-
-                /** Value nullValue. */
-                public nullValue: mesg.protobuf.NullValue;
-
-                /** Value numberValue. */
-                public numberValue: number;
-
-                /** Value stringValue. */
-                public stringValue: string;
-
-                /** Value boolValue. */
-                public boolValue: boolean;
-
-                /** Value structValue. */
-                public structValue?: (mesg.protobuf.IStruct|null);
-
-                /** Value listValue. */
-                public listValue?: (mesg.protobuf.IListValue|null);
-
-                /** Value kind. */
-                public kind?: ("nullValue"|"numberValue"|"stringValue"|"boolValue"|"structValue"|"listValue");
-            }
-
-            /** NullValue enum. */
-            enum NullValue {
-                NULL_VALUE = 0
-            }
-
-            /** Properties of a ListValue. */
-            interface IListValue {
-
-                /** ListValue values */
-                values?: (mesg.protobuf.IValue[]|null);
-            }
-
-            /** Represents a ListValue. */
-            class ListValue implements IListValue {
-
-                /**
-                 * Constructs a new ListValue.
-                 * @param [properties] Properties to set
-                 */
-                constructor(properties?: mesg.protobuf.IListValue);
-
-                /** ListValue values. */
-                public values: mesg.protobuf.IValue[];
             }
         }
 
@@ -594,8 +575,8 @@ declare namespace mesg {
             /** Properties of a CreateProcessRequest. */
             interface ICreateProcessRequest {
 
-                /** CreateProcessRequest key */
-                key?: (string|null);
+                /** CreateProcessRequest name */
+                name?: (string|null);
 
                 /** CreateProcessRequest nodes */
                 nodes?: (mesg.types.Process.INode[]|null);
@@ -613,8 +594,8 @@ declare namespace mesg {
                  */
                 constructor(properties?: mesg.api.ICreateProcessRequest);
 
-                /** CreateProcessRequest key. */
-                public key: string;
+                /** CreateProcessRequest name. */
+                public name: string;
 
                 /** CreateProcessRequest nodes. */
                 public nodes: mesg.types.Process.INode[];