+++ packages/api/src/types.ts
@@ -1,6 +1,5 @@
-import { AccountGetInputs, AccountGetOutputs, AccountListInputs, AccountListOutputs, AccountCreateInputs, AccountCreateOutputs, AccountDeleteInputs, AccountDeleteOutputs } from './account'
 import { EventCreateInputs, EventCreateOutputs, EventStreamInputs, EventStreamOutputs } from './event'
-import { ExecutionGetInputs, ExecutionGetOutputs, ExecutionStreamInputs, ExecutionStreamOutputs, ExecutionCreateInputs, ExecutionCreateOutputs, ExecutionUpdateInputs, ExecutionUpdateOutputs } from './execution'
+import { ExecutionGetInputs, ExecutionGetOutputs, ExecutionStreamInputs, ExecutionStreamOutputs, ExecutionCreateInputs, ExecutionCreateOutputs, ExecutionUpdateInputs, ExecutionUpdateOutputs, ExecutionListInputs, ExecutionListOutputs } from './execution'
 import { InstanceGetInputs, InstanceGetOutputs, InstanceListInputs, InstanceListOutputs } from './instance'
 import { RunnerGetInputs, RunnerGetOutputs, RunnerListInputs, RunnerListOutputs, RunnerCreateInputs, RunnerCreateOutputs, RunnerDeleteInputs, RunnerDeleteOutputs } from './runner'
 import { ServiceGetInputs, ServiceGetOutputs, ServiceHashInputs, ServiceHashOutputs, ServiceExistsInputs, ServiceExistsOutputs, ServiceListInputs, ServiceListOutputs, ServiceCreateInputs, ServiceCreateOutputs } from './service'
@@ -18,12 +17,6 @@ export const ExecutionStatus = {
 }
 
 export type IApi = {
-  account: {
-    get: (request: AccountGetInputs) => AccountGetOutputs
-    list: (request: AccountListInputs) => AccountListOutputs
-    create: (request: AccountCreateInputs) => AccountCreateOutputs
-    delete: (request: AccountDeleteInputs) => AccountDeleteOutputs
-  },
   event: {
     create: (request: EventCreateInputs) => EventCreateOutputs
     stream: (request: EventStreamInputs) => EventStreamOutputs
@@ -33,6 +26,7 @@ export type IApi = {
     stream: (request: ExecutionStreamInputs) => ExecutionStreamOutputs
     create: (request: ExecutionCreateInputs) => ExecutionCreateOutputs
     update: (request: ExecutionUpdateInputs) => ExecutionUpdateOutputs
+    list: (request: ExecutionListInputs) => ExecutionListOutputs
   }
   instance: {
     get: (request: InstanceGetInputs) => InstanceGetOutputs