+++ packages/cli/src/commands/daemon/logs.ts
@@ -10,7 +10,7 @@ export default class Logs extends Command {
     ...Command.flags,
     tail: flags.integer({
       description:  'Display the last N lines',
-      default: -1
+      default: 10000
     }),
     follow: flags.boolean({
       description: 'Follow logs',
@@ -24,7 +24,9 @@ export default class Logs extends Command {
     const services = await this.listServices({
       name: flags.name
     })
-    if (services.length === 0) return
+    if (services.length === 0) {
+      throw new Error("No engine is running.")
+    }
     const service = services[0]
     const logs: any = await service.logs({
       stderr: true,