+++ packages/cli/CHANGELOG.md
@@ -1,11 +1,31 @@
 # Changelog
 
-## [Unreleased](https://github.com/mesg-foundation/cli/releases/tag/vX.X.X)
+## [Unreleased](https://github.com/mesg-foundation/js-sdk/releases/tag/%40mesg%2Fcli%40X.X.X)
 
 #### Breaking Changes
-#### Added
-#### Changed
-#### Fixed
-#### Removed
+#### Improvements
+#### Bug fixes
+
+## [v0.2.0](https://github.com/mesg-foundation/js-sdk/releases/tag/%40mesg%2Fcli%400.2.0)
+
+#### Improvements
+
+- [#168](https://github.com/mesg-foundation/js-sdk/pull/168) Fetch and configure the engine to start on a specific network.
+
+## [v0.1.3](https://github.com/mesg-foundation/js-sdk/releases/tag/%40mesg%2Fcli%400.1.3)
+
+#### Bug fixes
+
+- [#163](https://github.com/mesg-foundation/js-sdk/pull/163) Fixes encoding of hash in not existing task error in command `service:execute`.
+
+## [v0.1.2](https://github.com/mesg-foundation/js-sdk/releases/tag/%40mesg%2Fcli%400.1.2)
+
+- [#161](https://github.com/mesg-foundation/js-sdk/pull/161) Use Engine v0.18.
+
+## [v0.1.1](https://github.com/mesg-foundation/js-sdk/releases/tag/%40mesg%2Fcli%400.1.1)
+
+#### Improvements
+
+- [#148](https://github.com/mesg-foundation/js-sdk/pull/148) Update process's compilation based on the latest API
 
 This library has been moved from https://github.com/mesg-foundation/cli, please check https://github.com/mesg-foundation/cli/blob/master/CHANGELOG.md for previous updates